x/cctp/types/message_pause_burning_and_minting_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/strangelove-ventures/noble/testutil/sample"
	"github.com/stretchr/testify/require"
)

func TestMsgPauseBurningAndMinting_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgPauseBurningAndMinting
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgPauseBurningAndMinting{
				From: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgPauseBurningAndMinting{
				From: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
